requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.1.2
python-dotenv==1.0.0
pydantic-settings==2.1.0
email-validator==2.1.0
werkzeug==3.0.1
pyjwt==2.8.0
faiss-cpu==1.7.4
numpy==1.24.3
sentence-transformers==2.2.2